+++ src/errors.ts
@@ -14,28 +14,22 @@ export class DependsOnMethodError extends RoutingError {
 
 /**
  * @desc An error related to the generating of the documentation
- * @todo rename to DocumentationError in v11
- * @todo remove params:string variation in v11
  * */
-export class OpenAPIError extends Error {
-  public override name = "OpenAPIError";
+export class DocumentationError extends Error {
+  public override name = "DocumentationError";
 
-  constructor(
-    params:
-      | string // backward compatibility
-      | ({ message: string } & Pick<
-          OpenAPIContext,
-          "path" | "method" | "isResponse"
-        >)
-  ) {
-    const finalMessage =
-      typeof params === "string"
-        ? params
-        : `${params.message}\nCaused by ${
-            params.isResponse ? "response" : "input"
-          } schema of an Endpoint assigned to ${params.method.toUpperCase()} method of ${
-            params.path
-          } path.`;
+  constructor({
+    message,
+    method,
+    path,
+    isResponse,
+  }: { message: string } & Pick<
+    OpenAPIContext,
+    "path" | "method" | "isResponse"
+  >) {
+    const finalMessage = `${message}\nCaused by ${
+      isResponse ? "response" : "input"
+    } schema of an Endpoint assigned to ${method.toUpperCase()} method of ${path} path.`;
     super(finalMessage);
   }
 }