+++ tests/unit/integration.spec.ts
@@ -1,6 +1,6 @@
 import { z } from "zod";
 import { routing } from "../../example/routing";
-import { Client, Integration, defaultEndpointsFactory } from "../../src";
+import { Integration, defaultEndpointsFactory } from "../../src";
 
 describe("API Integration Generator", () => {
   test.each(["client", "types"] as const)(
@@ -54,12 +54,4 @@ describe("API Integration Generator", () => {
       expect(client.print()).toMatchSnapshot();
     }
   );
-});
-
-describe("Client generator", () => {
-  test("Should be the same as Integration with the client variant", () => {
-    const gauge = new Integration({ variant: "client", routing }).print();
-    const client = new Client({ routing });
-    expect(client.print()).toEqual(gauge);
-  });
 });