+++ CHANGELOG.md
@@ -1,6 +1,20 @@
 # Changelog
 
+## Version 11
+
+- **Breaking changes**:
+  - Minimum Node version supported: 16.14.0.
+  - `OpenAPIError` renamed to `DocumentationError`.
+    - It's also now only accepts object argument. Use its `message` prop instead.
+  - `OpenAPI` class removed. Use `Documentation` one instead (same constructor props).
+  - `Client` class removed. Use `Integration` one instead.
+    - Instantiate by `new Integration({ variant: "client" })` for achieving the same behavior.
+
 ## Version 10
+
+### v10.9.0
+
+- `winston` version is 3.9.0.
 
 ### v10.8.1